# docs/generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/customer/home": {
      "get": {
        "description": "REGISTERED_USER only. Returns safe marketplace categories, discounted gifts, default address, and upcoming reminder.",
        "operationId": "CustomerMarketplaceController_home",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Customer home fetched successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch customer app home",
        "tags": [
          "Customer Marketplace"
        ]
      }
    },
    "/api/v1/customer/categories": {
      "get": {
        "description": "REGISTERED_USER only. Returns active categories that have available approved gifts.",
        "operationId": "CustomerMarketplaceController_categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Customer categories fetched successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List customer marketplace categories",
        "tags": [
          "Customer Marketplace"
        ]
      }
    },
    "/api/v1/customer/gifts/discounted": {
      "get": {
        "description": "REGISTERED_USER only. Reuses marketplace gift filters with offerOnly=true.",
        "operationId": "CustomerMarketplaceController_discountedGifts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "perfume",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "cmf0giftcatperfumes001",
              "type": "string"
            }
          },
          {
            "name": "categorySlug",
            "required": false,
            "in": "query",
            "schema": {
              "example": "perfumes",
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "cmf0providerflorist001",
              "type": "string"
            }
          },
          {
            "name": "offerOnly",
            "required": false,
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "example": 250,
              "type": "number"
            }
          },
          {
            "name": "minRating",
            "required": false,
            "in": "query",
            "schema": {
              "example": 4,
              "type": "number"
            }
          },
          {
            "name": "brand",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Dcodax Gifts",
              "type": "string"
            }
          },
          {
            "name": "deliveryOption",
            "required": false,
            "in": "query",
            "schema": {
              "example": "SAME_DAY",
              "type": "string",
              "enum": [
                "SAME_DAY",
                "NEXT_DAY",
                "SCHEDULED"
              ]
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "example": "popularity",
              "type": "string",
              "enum": [
                "popularity",
                "priceLowToHigh",
                "priceHighToLow",
                "rating",
                "newest",
                "discount"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discounted gifts fetched successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List discounted customer gifts",
        "tags": [
          "Customer Marketplace"
        ]
      }
    },
    "/api/v1/customer/gifts/filter-options": {
      "get": {
        "description": "REGISTERED_USER only. Brands are derived from approved active provider business names.",
        "operationId": "CustomerMarketplaceController_filterOptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Gift filter options fetched successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch marketplace gift filter options",
        "tags": [
          "Customer Marketplace"
        ]
      }
    },
    "/api/v1/customer/gifts": {
      "get": {
        "description": "REGISTERED_USER only. Only approved, published, active, in-stock gifts from approved active providers are returned. Active offers are calculated by the backend.",
        "operationId": "CustomerMarketplaceController_gifts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "perfume",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "cmf0giftcatperfumes001",
              "type": "string"
            }
          },
          {
            "name": "categorySlug",
            "required": false,
            "in": "query",
            "schema": {
              "example": "perfumes",
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "cmf0providerflorist001",
              "type": "string"
            }
          },
          {
            "name": "offerOnly",
            "required": false,
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "example": 250,
              "type": "number"
            }
          },
          {
            "name": "minRating",
            "required": false,
            "in": "query",
            "schema": {
              "example": 4,
              "type": "number"
            }
          },
          {
            "name": "brand",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Dcodax Gifts",
              "type": "string"
            }
          },
          {
            "name": "deliveryOption",
            "required": false,
            "in": "query",
            "schema": {
              "example": "SAME_DAY",
              "type": "string",
              "enum": [
                "SAME_DAY",
                "NEXT_DAY",
                "SCHEDULED"
              ]
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "example": "popularity",
              "type": "string",
              "enum": [
                "popularity",
                "priceLowToHigh",
                "priceHighToLow",
                "rating",
                "newest",
                "discount"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer gifts fetched successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List customer marketplace gifts",
        "tags": [
          "Customer Marketplace"
        ]
      }
    },
    "/api/v1/customer/gifts/{id}": {
      "get": {
        "description": "REGISTERED_USER only. Hidden/admin-only gift records are never returned.",
        "operationId": "CustomerMarketplaceController_giftDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Gift details fetched successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch customer-safe gift details",
        "tags": [
          "Customer Marketplace"
        ]
      }
    },
    "/api/v1/customer/wishlist": {
      "get": {
        "description": "REGISTERED_USER only. Returns customer-safe available gifts.",
        "operationId": "CustomerMarketplaceController_wishlist",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List wishlist gifts",
        "tags": [
          "Customer Wishlist"
        ]
      }
    },
    "/api/v1/customer/wishlist/{giftId}": {
      "post": {
        "description": "REGISTERED_USER only. Gift must be active, approved, published, and in stock. Duplicate wishlist entries are ignored.",
        "operationId": "CustomerMarketplaceController_addWishlist",
        "parameters": [
          {
            "name": "giftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add gift to wishlist",
        "tags": [
          "Customer Wishlist"
        ]
      },
      "delete": {
        "description": "REGISTERED_USER only. Removes only the current customer wishlist row.",
        "operationId": "CustomerMarketplaceController_removeWishlist",
        "parameters": [
          {
            "name": "giftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove gift from wishlist",
        "tags": [
          "Customer Wishlist"
        ]
      }
    },
    "/api/v1/customer/addresses": {
      "get": {
        "description": "REGISTERED_USER only. Customers can only view their own non-deleted addresses.",
        "operationId": "CustomerMarketplaceController_addresses",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List customer addresses",
        "tags": [
          "Customer Addresses"
        ]
      },
      "post": {
        "description": "REGISTERED_USER only. Maintains one default address per customer.",
        "operationId": "CustomerMarketplaceController_createAddress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create customer address",
        "tags": [
          "Customer Addresses"
        ]
      }
    },
    "/api/v1/customer/addresses/{id}": {
      "get": {
        "description": "REGISTERED_USER only. Address must belong to the current customer.",
        "operationId": "CustomerMarketplaceController_addressDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch customer address",
        "tags": [
          "Customer Addresses"
        ]
      },
      "patch": {
        "description": "REGISTERED_USER only. Maintains one default address per customer.",
        "operationId": "CustomerMarketplaceController_updateAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update customer address",
        "tags": [
          "Customer Addresses"
        ]
      },
      "delete": {
        "description": "REGISTERED_USER only. Address is soft deleted and removed from default status.",
        "operationId": "CustomerMarketplaceController_deleteAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete customer address",
        "tags": [
          "Customer Addresses"
        ]
      }
    },
    "/api/v1/customer/addresses/{id}/default": {
      "patch": {
        "description": "REGISTERED_USER only. Clears default flag from all other customer addresses.",
        "operationId": "CustomerMarketplaceController_setDefaultAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set default customer address",
        "tags": [
          "Customer Addresses"
        ]
      }
    },
    "/api/v1/customer/cart": {
      "get": {
        "description": "REGISTERED_USER only. Totals are backend calculated from price snapshots.",
        "operationId": "CustomerMarketplaceController_cart",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch active cart",
        "tags": [
          "Customer Cart"
        ]
      },
      "delete": {
        "description": "REGISTERED_USER only. Removes all items from active cart.",
        "operationId": "CustomerMarketplaceController_clearCart",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Clear active cart",
        "tags": [
          "Customer Cart"
        ]
      }
    },
    "/api/v1/customer/cart/items": {
      "post": {
        "description": "REGISTERED_USER only. Backend calculates unit price, active offer discount, final price, subtotal, delivery fee placeholder, and total.",
        "operationId": "CustomerMarketplaceController_addCartItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              },
              "examples": {
                "sendGift": {
                  "value": {
                    "giftId": "cmf0giftroses001",
                    "variantId": "cmf0variant50ml001",
                    "quantity": 1,
                    "deliveryOption": "SAME_DAY",
                    "recipientContactId": "cmf0contactmary001",
                    "recipientName": "Sarah Khan",
                    "recipientPhone": "+923001234567",
                    "recipientAddressId": "cmf0addresshome001",
                    "eventId": "cmf0eventbirthday001",
                    "giftMessage": "Hope you love this special surprise!",
                    "messageMediaUrls": [
                      "https://cdn.yourdomain.com/gift-message-media/photo.png"
                    ],
                    "scheduledDeliveryAt": "2026-12-24T10:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add item to cart",
        "tags": [
          "Customer Cart"
        ]
      }
    },
...
    "/api/v1/customer/orders/{id}": {
      "get": {
        "description": "REGISTERED_USER only. Order must belong to the current customer.",
        "operationId": "CustomerMarketplaceController_orderDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch customer order",
        "tags": [
          "Customer Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gift App Backend API",
    "description": "Gift App authentication and user account APIs",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCustomerAddressDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "Home"
          },
          "fullName": {
            "type": "string",
            "example": "Sarah Khan"
          },
          "phone": {
            "type": "string",
            "example": "+923001234567"
          },
          "line1": {
            "type": "string",
            "example": "House 12, Street 4, F-8/2"
          },
          "line2": {
            "type": "string",
            "example": "Near Centaurus Mall"
          },
          "city": {
            "type": "string",
            "example": "Islamabad"
          },
          "state": {
            "type": "string",
            "example": "Islamabad Capital Territory"
          },
          "country": {
            "type": "string",
            "example": "Pakistan"
          },
          "postalCode": {
            "type": "string",
            "example": "44000"
          },
          "latitude": {
            "type": "number",
            "example": 33.6844
          },
          "longitude": {
            "type": "number",
            "example": 73.0479
          },
          "deliveryInstructions": {
            "type": "string",
            "example": "Leave at reception."
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "label",
          "fullName",
          "phone",
          "line1",
          "city",
          "country"
        ]
      },
      "UpdateCustomerAddressDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "Home"
          },
          "fullName": {
            "type": "string",
            "example": "Sarah Khan"
          },
          "phone": {
            "type": "string",
            "example": "+923001234567"
          },
          "line1": {
            "type": "string",
            "example": "House 12, Street 4, F-8/2"
          },
          "line2": {
            "type": "string",
            "example": "Near Centaurus Mall"
          },
          "city": {
            "type": "string",
            "example": "Islamabad"
          },
          "state": {
            "type": "string",
            "example": "Islamabad Capital Territory"
          },
          "country": {
            "type": "string",
            "example": "Pakistan"
          },
          "postalCode": {
            "type": "string",
            "example": "44000"
          },
          "latitude": {
            "type": "number",
            "example": 33.6844
          },
          "longitude": {
            "type": "number",
            "example": 73.0479
          },
          "deliveryInstructions": {
            "type": "string",
            "example": "Leave at reception."
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}